requirements.txt
uuid
google-generativeai
pyautogui
questionary
python-dotenv
eel
pyaudio
numpy
Pillow
pywebview
